Mikudeptrai

+++ admin-dashboard/src/api/courseAPI.js
@@ -13,7 +13,7 @@ export const courseService = {
   // Get all courses
   getAllCourses: async () => {
     try {
-      const response = await api.get("/api/courses");
+      const response = await api.get("/api/admin/courses");
       console.log(response.data);
 
       return response.data.body || [];
@@ -27,7 +27,7 @@ export const courseService = {
   // // Get course by ID
   // getCourseById: async (id) => {
   //   try {
-  //     const response = await api.get(`/api/courses/${id}`);
+  //     const response = await api.get(`/api/admin/courses/${id}`);
   //     return response.data.body;
   //   } catch (error) {
   //     console.error(`Error fetching course with id ${id}:`, error);
@@ -38,7 +38,7 @@ export const courseService = {
   // Create new course
   createCourse: async (courseData) => {
     try {
-      const response = await api.post("/api/courses", courseData);
+      const response = await api.post("/api/admin/courses", courseData);
       return response.data;
     } catch (error) {
       console.error("Error creating course:", error);
@@ -49,7 +49,7 @@ export const courseService = {
   // Update course
   updateCourse: async (id, courseData) => {
     try {
-      const response = await api.put(`/api/courses/${id}`, courseData);
+      const response = await api.put(`/api/admin/courses/${id}`, courseData);
       return response.data;
     } catch (error) {
       console.error(`Error updating course with id ${id}:`, error);
@@ -60,7 +60,7 @@ export const courseService = {
   // Delete course
   deleteCourse: async (id) => {
     try {
-      const response = await api.delete(`/api/courses/${id}`);
+      const response = await api.delete(`/api/admin/courses/${id}`);
       return response.data;
     } catch (error) {
       console.error(`Error deleting course with id ${id}:`, error);
@@ -71,9 +71,7 @@ export const courseService = {
   // Toggle course visibility
   toggleCourseVisibility: async (id, isEnabled) => {
     try {
-      const response = await api.patch(`/api/courses/${id}/visibility`, {
-        enabled: isEnabled,
-      });
+      const response = await api.patch(`/api/admin/courses/${id}/visibility?isEnabled=${isEnabled}`);
       return response.data;
     } catch (error) {
       console.error(
